re-implement the usecase and its implementation
